Remove run completion logic
