+++ hoomd/Analyzer.cc
@@ -43,7 +43,6 @@ void export_Analyzer(py::module& m)
         .def(py::init< std::shared_ptr<SystemDefinition> >())
         .def("analyze", &Analyzer::analyze)
         .def("setProfiler", &Analyzer::setProfiler)
-        .def("isComplete", &Analyzer::isComplete)
         .def("notifyDetach", &Analyzer::notifyDetach)
     #ifdef ENABLE_MPI
         .def("setCommunicator", &Analyzer::setCommunicator)
