+++ hoomd/Updater.cc
@@ -42,7 +42,6 @@ void export_Updater(py::module& m)
     .def("update", &Updater::update)
     .def("setProfiler", &Updater::setProfiler)
     .def("notifyDetach", &Updater::notifyDetach)
-    .def("isComplete", &Updater::isComplete)
     #ifdef ENABLE_MPI
     .def("setCommunicator", &Updater::setCommunicator)
     #endif
